+++ internal/js_parser/js_parser.go
@@ -14887,8 +14887,8 @@ func Parse(log logger.Log, source logger.Source, options Options) (result js_ast
 		// silently changed in TypeScript 4.3. It's a breaking change even though
 		// it wasn't mentioned in the announcement blog post for TypeScript 4.3:
 		// https://devblogs.microsoft.com/typescript/announcing-typescript-4-3/.
-		if options.targetFromAPI == config.TargetWasConfiguredIncludingESNext ||
-			(options.tsTarget != nil && strings.EqualFold(options.tsTarget.Target, "ESNext")) {
+		if options.targetFromAPI == config.TargetWasConfiguredAndAtLeastES2022 ||
+			(options.tsTarget != nil && options.tsTarget.TargetIsAtLeastES2022) {
 			options.useDefineForClassFields = config.True
 		} else {
 			options.useDefineForClassFields = config.False