+++ pkg/api/api_impl.go
@@ -309,8 +309,9 @@ func validateFeatures(log logger.Log, target Target, engines []Engine) (config.T
 		constraints[compat.ES] = []int{2021}
 	case ES2022:
 		constraints[compat.ES] = []int{2022}
+		targetFromAPI = config.TargetWasConfiguredAndAtLeastES2022
 	case ESNext:
-		targetFromAPI = config.TargetWasConfiguredIncludingESNext
+		targetFromAPI = config.TargetWasConfiguredAndAtLeastES2022
 	case DefaultTarget:
 	default:
 		panic("Invalid target")